+++ packages/signals_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- `Watch` widget improvements (reducing extra first frame build)
+
 ## 1.0.0
 
 - Moving core api from `signals` package to `signals_core` package